+++ src/main/java/com/example/uambite/dto/response/IngredienteExtraResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -14,5 +15,5 @@ import java.util.UUID;
 public class IngredienteExtraResponse {
     private UUID id;
     private String nombre;
-    private Double precioExtra;
+    private BigDecimal precioExtra;
 }
