+++ src/main/java/com/example/uambite/dto/response/PedidoResponse.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -18,9 +19,9 @@ import java.util.UUID;
 public class PedidoResponse {
     private UUID id;
     private EstadoPedido estado;
-    private Double total;
-    private Double subtotal;
-    private Double descuentoAplicado;
+    private BigDecimal total;
+    private BigDecimal subtotal;
+    private BigDecimal descuentoAplicado;
     private TipoEntrega tipoEntrega;
     private UUID usuarioId;
     private String usuario;
